Introduce Stack.Cache and Stack.CacheService
Add caching so that services can be fetched by key so that do not
need to pass the service datastructure around. Also cache the
macro escapable data structures using a compiled module so that as
much as possible is in the literal pool to avoid copying.

The cached function is a local fun so that arguments can be
versioned with the matching module version, so the previous
cached version will still work (but prior ones will fail with
bad function exception). The local fun also does a tail call
immediately so is on the call stack for minimal amount of time.